Window demo 2 + removed GLStateManager

+++ BleachHack-Fabric/src/main/java/bleach/hack/gui/NotebotScreen.java
@@ -59,77 +59,84 @@ public class NotebotScreen extends AbstractWindowScreen {
 		renderBackground();
 		super.render(int_1, int_2, float_1);
 		
-		int x = windows.get(0).x1,
-				y = windows.get(0).y1 + 10,
-				w = width / 2,
-				h = height / 2;
+		renderWindow(0, int_1, int_2);
 		
-		drawCenteredString(font, "Tutorial..", x + w - 24, y + 4, 0x9090c0);
-		
-		int pageEntries = 0;
-		for(int i = y + 20; i < y + h - 27; i += 10) pageEntries++;
-		
-		drawCenteredString(font, "<  Page " + (page + 1) + "  >", x + 55, y + 5, 0xc0c0ff);
-		
-		fillButton(x + 10, y + h - 13, x + 99, y + h - 3, 0xff3a3a3a, 0xff353535, int_1, int_2);
-		drawCenteredString(font, "Download Songs..", x + 55, y + h - 12, 0xc0dfdf);
-		
-		int c = 0, c1 = -1;
-		for(String s: files) {
-			c1++;
-			if(c1 < page * pageEntries) continue;
-			if(c1 > (page + 1) * pageEntries) break;
+		if(!windows.get(0).closed) {
+			int x = windows.get(0).x1,
+					y = windows.get(0).y1 + 10,
+					w = width / 2,
+					h = height / 2;
 			
-			fillButton(x + 5, y + 15 + c * 10, x + 105, y + 25 + c * 10, 
-					Notebot.filePath.equals(s) ? 0xf0408040 : selected.equals(s) ? 0xf0202020 : 0xf0404040, 0xf0303030, int_1, int_2);
-			if(cutText(s, 105).equals(s)) drawCenteredString(font, s, x + 55, y + 16 + c * 10, -1);
-			else drawString(font, cutText(s, 105), x + 5, y + 16 + c * 10, -1);
-			c++;
-		}
-		
-		if(entry != null) {
-			drawCenteredString(font, entry.fileName, x + w - w / 4, y + 10, 0xa030a0);
-			drawCenteredString(font, entry.length / 20 + "s", x + w - w / 4, y + 20, 0xc000c0);
-			drawCenteredString(font, "Notes: ", x + w - w / 4, y + 38, 0x80f080);
+			drawCenteredString(font, "Tutorial..", x + w - 24, y + 4, 0x9090c0);
 			
-			int c2 = 0;
-			for(Entry<Instrument, Integer> e: entry.notes.entrySet()) {
-				itemRenderer.zOffset = 500 - c2 * 20;
-				drawCenteredString(font, StringUtils.capitalize(e.getKey().asString()) + " x" + e.getValue(), 
-						x + w - w / 4, y + 50 + c2 * 10, 0x50f050);
-				GL11.glPushMatrix();
-				GuiLighting.enableForItems();
-				if(e.getKey() == Instrument.HARP) itemRenderer.renderGuiItem(new ItemStack(Items.DIRT), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.BASEDRUM) itemRenderer.renderGuiItem(new ItemStack(Items.STONE), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.SNARE) itemRenderer.renderGuiItem(new ItemStack(Items.SAND), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.HAT) itemRenderer.renderGuiItem(new ItemStack(Items.GLASS), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.BASS) itemRenderer.renderGuiItem(new ItemStack(Items.OAK_WOOD), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.FLUTE) itemRenderer.renderGuiItem(new ItemStack(Items.CLAY), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.BELL) itemRenderer.renderGuiItem(new ItemStack(Items.GOLD_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.GUITAR) itemRenderer.renderGuiItem(new ItemStack(Items.WHITE_WOOL), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.CHIME) itemRenderer.renderGuiItem(new ItemStack(Items.PACKED_ICE), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.XYLOPHONE) itemRenderer.renderGuiItem(new ItemStack(Items.BONE_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.IRON_XYLOPHONE) itemRenderer.renderGuiItem(new ItemStack(Items.IRON_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.COW_BELL) itemRenderer.renderGuiItem(new ItemStack(Items.SOUL_SAND), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.DIDGERIDOO) itemRenderer.renderGuiItem(new ItemStack(Items.PUMPKIN), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.BIT) itemRenderer.renderGuiItem(new ItemStack(Items.EMERALD_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.BANJO) itemRenderer.renderGuiItem(new ItemStack(Items.HAY_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				if(e.getKey() == Instrument.PLING) itemRenderer.renderGuiItem(new ItemStack(Items.GLOWSTONE), x + w - w / 4 + 40, y + 46 + c2 * 10);
-				c2++;
-				GL11.glPopMatrix();
+			int pageEntries = 0;
+			for(int i = y + 20; i < y + h - 27; i += 10) pageEntries++;
+			
+			drawCenteredString(font, "<  Page " + (page + 1) + "  >", x + 55, y + 5, 0xc0c0ff);
+			
+			fillButton(x + 10, y + h - 13, x + 99, y + h - 3, 0xff3a3a3a, 0xff353535, int_1, int_2);
+			drawCenteredString(font, "Download Songs..", x + 55, y + h - 12, 0xc0dfdf);
+			
+			int c = 0, c1 = -1;
+			for(String s: files) {
+				c1++;
+				if(c1 < page * pageEntries) continue;
+				if(c1 > (page + 1) * pageEntries) break;
+				
+				fillButton(x + 5, y + 15 + c * 10, x + 105, y + 25 + c * 10, 
+						Notebot.filePath.equals(s) ? 0xf0408040 : selected.equals(s) ? 0xf0202020 : 0xf0404040, 0xf0303030, int_1, int_2);
+				if(cutText(s, 105).equals(s)) drawCenteredString(font, s, x + 55, y + 16 + c * 10, -1);
+				else drawString(font, cutText(s, 105), x + 5, y + 16 + c * 10, -1);
+				c++;
 			}
 			
-			fillButton(x + w - w / 2 + 10, y + h - 15, x + w - w / 4, y + h - 5, 0xff903030, 0xff802020, int_1, int_2);
-			fillButton(x + w - w / 4 + 5, y + h - 15, x + w - 5, y + h - 5, 0xff308030, 0xff207020, int_1, int_2);
-			fillButton(x + w - w / 4 - w / 8, y + h - 27, x + w - w / 4 + w / 8, y + h - 17, 0xff303080, 0xff202070, int_1, int_2);
-			
-			int pixels = (int) Math.round(MathHelper.clamp((w / 4)*((double)entry.playTick / (double)entry.length), 0, w / 4));
-			fill(x + w - w / 4 - w / 8, y + h - 27, (x + w - w / 4 - w / 8) + pixels, y + h - 17, 0x507050ff);
-			
-			drawCenteredString(font, "Delete", (int)(x + w - w / 2.8), y + h - 14, 0xff0000);
-			drawCenteredString(font, "Select", x + w - w / 8, y + h - 14, 0x00ff00);
-			drawCenteredString(font, (entry.playing ? "Playing" : "Play") + " (scuffed)", x + w - w / 4, y + h - 26, 0x6060ff);
+			if(entry != null) {
+				drawCenteredString(font, entry.fileName, x + w - w / 4, y + 10, 0xa030a0);
+				drawCenteredString(font, entry.length / 20 + "s", x + w - w / 4, y + 20, 0xc000c0);
+				drawCenteredString(font, "Notes: ", x + w - w / 4, y + 38, 0x80f080);
+				
+				int c2 = 0;
+				for(Entry<Instrument, Integer> e: entry.notes.entrySet()) {
+					itemRenderer.zOffset = 500 - c2 * 20;
+					drawCenteredString(font, StringUtils.capitalize(e.getKey().asString()) + " x" + e.getValue(), 
+							x + w - w / 4, y + 50 + c2 * 10, 0x50f050);
+					GL11.glPushMatrix();
+					GuiLighting.enableForItems();
+					if(e.getKey() == Instrument.HARP) itemRenderer.renderGuiItem(new ItemStack(Items.DIRT), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.BASEDRUM) itemRenderer.renderGuiItem(new ItemStack(Items.STONE), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.SNARE) itemRenderer.renderGuiItem(new ItemStack(Items.SAND), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.HAT) itemRenderer.renderGuiItem(new ItemStack(Items.GLASS), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.BASS) itemRenderer.renderGuiItem(new ItemStack(Items.OAK_WOOD), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.FLUTE) itemRenderer.renderGuiItem(new ItemStack(Items.CLAY), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.BELL) itemRenderer.renderGuiItem(new ItemStack(Items.GOLD_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.GUITAR) itemRenderer.renderGuiItem(new ItemStack(Items.WHITE_WOOL), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.CHIME) itemRenderer.renderGuiItem(new ItemStack(Items.PACKED_ICE), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.XYLOPHONE) itemRenderer.renderGuiItem(new ItemStack(Items.BONE_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.IRON_XYLOPHONE) itemRenderer.renderGuiItem(new ItemStack(Items.IRON_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.COW_BELL) itemRenderer.renderGuiItem(new ItemStack(Items.SOUL_SAND), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.DIDGERIDOO) itemRenderer.renderGuiItem(new ItemStack(Items.PUMPKIN), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.BIT) itemRenderer.renderGuiItem(new ItemStack(Items.EMERALD_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.BANJO) itemRenderer.renderGuiItem(new ItemStack(Items.HAY_BLOCK), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					if(e.getKey() == Instrument.PLING) itemRenderer.renderGuiItem(new ItemStack(Items.GLOWSTONE), x + w - w / 4 + 40, y + 46 + c2 * 10);
+					c2++;
+					GL11.glPopMatrix();
+				}
+				
+				fillButton(x + w - w / 2 + 10, y + h - 15, x + w - w / 4, y + h - 5, 0xff903030, 0xff802020, int_1, int_2);
+				fillButton(x + w - w / 4 + 5, y + h - 15, x + w - 5, y + h - 5, 0xff308030, 0xff207020, int_1, int_2);
+				fillButton(x + w - w / 4 - w / 8, y + h - 27, x + w - w / 4 + w / 8, y + h - 17, 0xff303080, 0xff202070, int_1, int_2);
+				
+				int pixels = (int) Math.round(MathHelper.clamp((w / 4)*((double)entry.playTick / (double)entry.length), 0, w / 4));
+				fill(x + w - w / 4 - w / 8, y + h - 27, (x + w - w / 4 - w / 8) + pixels, y + h - 17, 0x507050ff);
+				
+				drawCenteredString(font, "Delete", (int)(x + w - w / 2.8), y + h - 14, 0xff0000);
+				drawCenteredString(font, "Select", x + w - w / 8, y + h - 14, 0x00ff00);
+				drawCenteredString(font, (entry.playing ? "Playing" : "Play") + " (scuffed)", x + w - w / 4, y + h - 26, 0x6060ff);
+			}
 		}
+		
+		renderWindow(1, int_1, int_2);
+		renderWindow(2, int_1, int_2);
 	}
 	
 	public void tick() {
@@ -142,47 +149,49 @@ public class NotebotScreen extends AbstractWindowScreen {
 	}
 	
 	public boolean mouseClicked(double double_1, double double_2, int int_1) {
-		int x = windows.get(0).x1,
-				y = windows.get(0).y1 + 10,
-				w = width / 2,
-				h = height / 2;
-		
-		if(double_1 > x + 20 && double_1 < x + 35 && double_2 > y + 5 && double_2 < y + 15) if(page > 0) page--;
-		if(double_1 > x + 77 && double_1 < x + 92 && double_2 > y + 5 && double_2 < y + 15) page++;
-		if(double_1 > x + w - 44 && double_1 < x + w && double_2 > y + 3 && double_2 < y + 15) {
-			try { SystemUtil.getOperatingSystem().open(new URI("https://www.youtube.com/watch?v=clT_aNvQedk")); }catch(Exception e) {}
-		}
-		if(double_1 > x + 10 && double_1 < x + 99 && double_2 > y + h - 13 && double_2 < y + h - 3) {
-			NotebotUtils.downloadSongs(true);
-		}
-		
-		if(entry != null) {
-			/* Pfft why use buttons when you can use meaningless rectangles with messy code */
-			if(double_1 > x + w - w / 2 + 10 && double_1 < x + w - w / 4 && double_2 > y + h - 15 && double_2 < y + h - 5) {
-				BleachFileMang.deleteFile("notebot", entry.fileName);
-				minecraft.openScreen(this);
+		if(!windows.get(0).closed) {
+			int x = windows.get(0).x1,
+					y = windows.get(0).y1 + 10,
+					w = width / 2,
+					h = height / 2;
+			
+			if(double_1 > x + 20 && double_1 < x + 35 && double_2 > y + 5 && double_2 < y + 15) if(page > 0) page--;
+			if(double_1 > x + 77 && double_1 < x + 92 && double_2 > y + 5 && double_2 < y + 15) page++;
+			if(double_1 > x + w - 44 && double_1 < x + w && double_2 > y + 3 && double_2 < y + 15) {
+				try { SystemUtil.getOperatingSystem().open(new URI("https://www.youtube.com/watch?v=clT_aNvQedk")); }catch(Exception e) {}
 			}
-			if(double_1 > x + w - w / 4 + 5 && double_1 < x + w - 5 && double_2 > y + h - 15 && double_2 < y + h - 5) {
-				Notebot.filePath = entry.fileName;
+			if(double_1 > x + 10 && double_1 < x + 99 && double_2 > y + h - 13 && double_2 < y + h - 3) {
+				NotebotUtils.downloadSongs(true);
 			}
-			if(double_1 > x + w - w / 4 - w / 8 && double_1 < x + w - w / 4 + w / 8 && double_2 > y + h - 27 && double_2 < y + h - 17) {
-				entry.playing = !entry.playing;
+			
+			if(entry != null) {
+				/* Pfft why use buttons when you can use meaningless rectangles with messy code */
+				if(double_1 > x + w - w / 2 + 10 && double_1 < x + w - w / 4 && double_2 > y + h - 15 && double_2 < y + h - 5) {
+					BleachFileMang.deleteFile("notebot", entry.fileName);
+					minecraft.openScreen(this);
+				}
+				if(double_1 > x + w - w / 4 + 5 && double_1 < x + w - 5 && double_2 > y + h - 15 && double_2 < y + h - 5) {
+					Notebot.filePath = entry.fileName;
+				}
+				if(double_1 > x + w - w / 4 - w / 8 && double_1 < x + w - w / 4 + w / 8 && double_2 > y + h - 27 && double_2 < y + h - 17) {
+					entry.playing = !entry.playing;
+				}
 			}
-		}
-		
-		int pageEntries = 0;
-		for(int i = y + 20; i < y + h - 27; i += 10) pageEntries++;
-		
-		int c = 0;
-		int c1 = -1;
-		for(String s: files) {
-			c1++;
-			if(c1 < page * pageEntries) continue;
-			if(double_1 > x + 5 && double_1 < x + 105 && double_2 > y + 15 + c * 10 && double_2 < y + 25 + c * 10) {
-				entry = new NotebotEntry(s);
-				selected = s;
+			
+			int pageEntries = 0;
+			for(int i = y + 20; i < y + h - 27; i += 10) pageEntries++;
+			
+			int c = 0;
+			int c1 = -1;
+			for(String s: files) {
+				c1++;
+				if(c1 < page * pageEntries) continue;
+				if(double_1 > x + 5 && double_1 < x + 105 && double_2 > y + 15 + c * 10 && double_2 < y + 25 + c * 10) {
+					entry = new NotebotEntry(s);
+					selected = s;
+				}
+				c++;
 			}
-			c++;
 		}
 		
 		return super.mouseClicked(double_1, double_2, int_1);
